PULPO-19 Cleanup should not fail if eb environment does not exist

+++ aws/deployment-pipeline/cleanup/codepipeline-cleanup/prereqs/eb/delete_eb_environment.sh
@@ -17,5 +17,19 @@ done
 
 set -x
 
-aws elasticbeanstalk terminate-environment \
-	--environment-name ${APPLICATION_NAME}-${ENVIRONMENT_SUFFIX}
+ENVIRONMENT_NAME=${APPLICATION_NAME}-${ENVIRONMENT_SUFFIX}
+
+if (aws elasticbeanstalk describe-environments --application-name ${APPLICATION_NAME} --environment-names ${ENVIRONMENT_NAME} | jq '(.Environments[] | select(.Status != "Terminated"))' | grep -q ${ENVIRONMENT_NAME}); then
+
+  echo "EB environment ${ENVIRONMENT_NAME} for application ${APPLICATION_NAME} exists"
+
+  echo "Deleting EB environment ${ENVIRONMENT_NAME}"
+
+  aws elasticbeanstalk terminate-environment \
+	--environment-name ${ENVIRONMENT_NAME}
+
+else
+
+  echo "EB environment ${ENVIRONMENT_NAME} for application ${APPLICATION_NAME} does not exist. Skipping terminate."
+
+fi
